refactor: enhance type hints and improve handling of optional values
This update improves type hinting across various modules and ensures better handling of optional values in the codebase. The following changes were made:

- Updated type hints in `ConvertEmptyStringsToNoneMiddleware` for better clarity.
- Enhanced type definitions in `GearService` and `GearSettingsService` to use more explicit types.
- Improved handling of optional return values in repository methods to avoid potential errors.
- Added necessary imports for type definitions in several files.

These changes enhance code readability and maintainability while ensuring type safety throughout the application.

+++ backend/app/modules/ai/utils/models_config.py
@@ -155,4 +155,4 @@ def calculate_cost(model_id: str, prompt_tokens: int, completion_tokens: int) ->
     input_cost = (prompt_tokens / 1_000_000) * model["cost_per_1m_input"]
     output_cost = (completion_tokens / 1_000_000) * model["cost_per_1m_output"]
 
-    return round(input_cost + output_cost, 6)
+    return float(round(input_cost + output_cost, 6))
